feishu_mirror/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pypdf>=5.1.0
python-docx>=1.1.2
python-pptx>=1.0.2
openpyxl>=3.1.5
requests>=2.32.3
yfinance>=0.2.40
matplotlib>=3.9.0
